Fix compilation with custom CFLAGS
With this small change it is possible to build flashrom again when
specifying custom CFLAGS/LDFLAGS from the make command line like.

  make CFLAGS="..." LDFLAGS="..."

I need to do this when building flashrom in a cross compiler environment
like buildroot for a foreign target.

Corresponding to flashrom svn r196 and coreboot v2 svn r3102.

Signed-off-by: Clark Rawlins <clark@bit63.org>
Acked-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>