Original v2 revision: 3619
This patch fixes support for the AT49F002N(T) chip in the flashrom tool.

It replaces the write function to one based on write_byte_program_jedec()
instead of write_page_write_jedec(), as this part does not support page
programming.
I have verified the NT variant to fully work now, and adjusted the test
status accordingly. The N variant *should* also work with this patch, but
remains untested.

Signed-off-by: Tim ter Laak <timl@scintilla.utwente.nl>
Acked-by: Peter Stuge <peter@stuge.se>


git-svn-id: svn://flashrom.org/flashrom/trunk@320 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1