Add board-enable code for the Shuttle AK38N.
FYI, this board can only decode 256 KB chips (not 512 KB ones) unfortunately.
The highest address line (A18) is not connected on this board.

The it8705f_write_enable() is kept generic enough so it can be reused for other
board-enables, possibly in the board_biostar_p4m80_m4() for example, but that
shouldn't be touched for now, unless someone can test the code.

Signed-off-by: Uwe Hermann <uwe@hermann-uwe.de>
Acked-by: Luc Verhaegen <libv@skynet.be>



git-svn-id: svn://coreboot.org/flashrom/trunk@616 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1