There is no need to tell people to install libz if flashrom doesn't need libz
So far, the only case where libz is needed is when a library (libpci)
pulls in libz and even then it only happens if libpci is available in a
static version only and said static version has libz requirements.

Check for libpci separately and don't require libz if it isn't needed.

Clarify the README.

Corresponding to flashrom svn r743.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Stefan Reinauer <stepan@coresystems.de>