Split hardware access, OS abstraction and chip drivers out of flash.h to get a better overview of what belongs where.
This patch is only the first step, but it hopefully will make working
with the code and especially porting to new platforms easier.

Subsequent patches should move #includes for the newly created files
hwaccess.h and chipdrivers.h from flash.h to the files which need them.
Programmers should live in a separate header file as well.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Sean Nelson <audiohacked@gmail.com>
Tested-by: Idwer Vollering <vidwer@gmail.com>


git-svn-id: svn://coreboot.org/flashrom/trunk@799 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1