JEDEC ID probing checks the parity of the vendor ID and verifies that the ID differs from the flash chip contents. Add the same feature to 82802AB ID probing.
This should reduce the number of lines we have to look at to determine
if we're missing a chip definition or if we need a board enable. Just
use grep on the log:
grep -v "parity violation"
To narrow it down further, try:
grep -v "id1 is normal flash content, id2 is normal flash content"
And of course you want to ignore the skipped probes:
grep -v "skipped"
The remaining lines are worth examining, and if those look bogus as
well, you can bet that we just need a board enable.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Michael Karcher <flashrom@mkarcher.dialup.fu-berlin.de>


git-svn-id: svn://coreboot.org/flashrom/trunk@971 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1