Add a udev rules file for flashrom.
This allows USB-based external programmers to be used by non-root users
(which are in the 'plugdev' group). The file is to be installed by the
distros into the proper place (not sure if this is distro-specific). On
Debian the file will end up in /etc/udev/rules.d/z60_flashrom.rules.

On some systems the 'plugdev' group might have to adapted to whatever
the respective distro uses.

The following devices are listed so far:

 - Amontec JTAGkey(2)
 - Buspirate
 - Dediprog SF100
 - DLP Design DLP-USB1232H
 - FTDI FT4232H Mini-Module

Signed-off-by: Uwe Hermann <uwe@hermann-uwe.de>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>



git-svn-id: svn://coreboot.org/flashrom/trunk@1110 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1