make Makefile's test programs safe(r)
we don't use -W or similarly strict compiler checks (yet), but
enabling its use is a good thing. if you add -W to the cflags
without this patch, detection of the compiler will fail with gcc 4.4
for example, because compiling of the test program will fail due to
a warning of unused arguments and -Werror. similarly the other
checks involving compiling test programs would fail.

Signed-off-by: Stefan Tauner <stefan.tauner@student.tuwien.ac.at>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>

git-svn-id: svn://coreboot.org/flashrom/trunk@1356 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1