Add board enable for ASUS P5GD2 Premium
This is based on Joshua's patch, but with a DMI pattern, contrary to
http://www.flashrom.org/pipermail/flashrom/2010-September/004745.html
Rationale: although it looks like all P5GD2/C* might use the same
board enable (intel_ich_gpio21_raise), we need to add the variants
individually as long as we don't have enough reports to make a
better guess. This also guarantees correct output of flashrom -L and -z
and on activation of the board enable - not like this:
http://www.flashrom.org/pipermail/flashrom/attachments/20100930/2286e5d1/attachment-0003.txt

Success report:
http://www.flashrom.org/pipermail/flashrom/2010-August/004555.html

Also, fix last commit by adding a missing comma.

Corresponding to flashrom svn r1411.

Tested-by: Roberto Malinverni
Signed-off-by: Joshua Roys <roysjosh@gmail.com>
Acked-by: Stefan Tauner <stefan.tauner@alumni.tuwien.ac.at>