Move the main() function from flashrom.c to cli_classic.c.
The file flashrom.c is part of libflashrom and should thus not contain a
main() function, that would break compilation of all frontends using
libflashrom. Also, cli_classic.c is the right place anyway, as it's the
main() of the CLI tool.

Rename the simple wrapper cli_classic() in cli_classic.c to main(), as
it's not really needed.

Signed-off-by: Uwe Hermann <uwe@hermann-uwe.de>
Acked-by: Stefan Tauner <stefan.tauner@student.tuwien.ac.at>



git-svn-id: svn://flashrom.org/flashrom/trunk@1417 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1