Warn if we skipped all blocks while writing.
Quick hack to print a warning if we skip all blocks in a write
process because they are already equal to the requested image. We
want something like this to make users aware... and some developers
who regularly fall for this too *coughcough*. There might be more
elegant solutions for this... patches welcome.

This does not work for erasing because we do that unconditionally.

Signed-off-by: Stefan Tauner <stefan.tauner@student.tuwien.ac.at>
Acked-by: Idwer Vollering <vidwer@gmail.com>

git-svn-id: svn://coreboot.org/flashrom/trunk@1680 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1