Makefile: Warn if user tries to compile for libpayload w/o crossgcc
While flashrom is not as picky on compilers as coreboot, there is still a
high probablilty of breakage when one combines libpayload and distribution
compilers. Print a warning if we detect that to give the daring user a hint
how to resolve the explosions potentially following below it.

Corresponding to flashrom svn r1740.

Signed-off-by: Stefan Tauner <stefan.tauner@alumni.tuwien.ac.at>
Acked-by: Stefan Tauner <stefan.tauner@alumni.tuwien.ac.at>