Makefile: Look for libftdi >= 1.0 instead of libftdi >= 0.2
Modern distros no longer provide libftdi 0.2, and as a result,
programmers that depend on libftdi must be disabled. This means that
popular programmers, such as ft2232_spi are now disabled in modern
distros. Luckily the libftdi API did not change significantly, so
for our purposes, we can simply use the newer headers and link to the
newer library.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>