linux_spi: Dynamically detect max buffer size
The max buffer size of the linux kernel's SPI device is a compile-time parameter.
Read it on initialization and use it to inform the max transfer size. We allow
up to 5 bytes for the SPI command. On master, SPI commands are always 4 bytes,
but 4ba support will raise that to 5. Choosing 5 makes sure this will keep
working when 4ba support is merged.

Signed-off-by: Keno Fischer <keno@juliacomputing.com>