linux_spi: also change maximum read speed
Previously only the write speed would be set by `spispeed` argument.
In this commit, read speed will also be set by that.

Tested on Widora NEO with mainline OpenWrt (Linux 4.9)
and software emulated SPI (spi-gpio). No errors but SPI
remains on about 1.5MHz. This is due to the performance
limitation of spi-gpio driver in the linux kernel. The
hardware on this board is too weak to run with a higher
SPI clock speed (CPU is MTK's MT7688, single MIPS24kc
core @ 580MHz).

As a result, this patch should be tested by others again
if you have a more powerful device.

Change-Id: If834ab4a94f5f591137e3dede84bbb9530b65ad0
Signed-off-by: Jackson Ming Hu <huming2207@gmail.com>